+++ package.json
@@ -1,11 +1,11 @@
 {
-	"name": "reactfolio",
+	"name": "portfolio",
 	"version": "1.1.0",
 	"description": "Personal portfolio template built using Reactjs",
 	"author": {
-		"name": "Tharindu N. Madhusanka",
-		"email": "mail@tharindu.dev",
-		"url": "https://tharindu.dev"
+		"name": "Ethan Gray",
+		"email": "ethgray02@gmail.com",
+		"url": "https://ethgray.com"
 	},
 	"license": "MIT",
 	"private": true,
@@ -31,7 +31,9 @@
 		"start": "react-scripts start",
 		"build": "react-scripts build",
 		"test": "react-scripts test",
-		"eject": "react-scripts eject"
+		"eject": "react-scripts eject",
+		"predeploy": "npm run build",
+		"deploy": "gh-pages -d build"
 	},
 	"eslintConfig": {
 		"extends": [
@@ -54,5 +56,9 @@
 	"prettier": {
 		"useTabs": true,
 		"tabWidth": 4
+	},
+	"homepage": "https://www.ethgray.com",
+	"devDependencies": {
+		"gh-pages": "^6.3.0"
 	}
 }
