--- src/components/bender/Bender.css
@@ -1,65 +1,0 @@
-body {
-  font-family: "Orbitron", sans-serif;
-  background-color: #1a202c;
-  margin: 0;
-}
-
-/* Custom Planet Express colors */
-.bg-planet-express-dark {
-  background-color: #2c3e50;
-}
-.text-planet-express-accent {
-  color: #e74c3c;
-} /* An orange-red accent */
-
-.btn-planet-express {
-  background-color: #e74c3c;
-  transition:
-    background-color 0.3s ease,
-    transform 0.2s ease,
-    opacity 0.3s;
-}
-
-.btn-planet-express:hover:not(:disabled) {
-  background-color: #c0392b;
-  transform: scale(1.05);
-}
-
-.quote-bubble {
-  position: relative;
-  background: #2c3e50;
-  border-radius: 0.4em;
-}
-
-.quote-bubble:after {
-  content: "";
-  position: absolute;
-  bottom: 0;
-  left: 50%;
-  width: 0;
-  height: 0;
-  border: 20px solid transparent;
-  border-top-color: #2c3e50;
-  border-bottom: 0;
-  border-left: 0;
-  margin-left: -10px;
-  margin-bottom: -20px;
-}
-
-.loader {
-  border: 4px solid #f3f3f3;
-  border-radius: 50%;
-  border-top: 4px solid #e74c3c;
-  width: 40px;
-  height: 40px;
-  animation: spin 1s linear infinite;
-}
-
-@keyframes spin {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
-}
