+++ src/pages/styles/Bender.css
@@ -1,0 +1,182 @@
+body {
+  font-family: "Orbitron", sans-serif;
+  background-color: #1a202c;
+  margin: 0;
+}
+
+/* Custom Planet Express colors */
+.bg-planet-express-dark {
+  background-color: #2c3e50;
+}
+.text-planet-express-accent {
+  color: #e74c3c;
+} /* An orange-red accent */
+
+.btn-planet-express {
+  background-color: #e74c3c;
+  transition:
+    background-color 0.3s ease,
+    transform 0.2s ease,
+    opacity 0.3s;
+}
+
+.btn-planet-express:hover:not(:disabled) {
+  background-color: #c0392b;
+  transform: scale(1.05);
+}
+
+.quote-bubble {
+  position: relative;
+  background: #2c3e50;
+  border-radius: 0.4em;
+}
+
+.quote-bubble:after {
+  content: "";
+  position: absolute;
+  bottom: 0;
+  left: 50%;
+  width: 0;
+  height: 0;
+  border: 20px solid transparent;
+  border-top-color: #2c3e50;
+  border-bottom: 0;
+  border-left: 0;
+  margin-left: -10px;
+  margin-bottom: -20px;
+}
+
+.loader {
+  border: 4px solid #f3f3f3;
+  border-radius: 50%;
+  border-top: 4px solid #e74c3c;
+  width: 40px;
+  height: 40px;
+  animation: spin 1s linear infinite;
+}
+
+@keyframes spin {
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(360deg);
+  }
+}
+
+.bender-main {
+  background-color: #2c3e50;
+  min-height: 100vh;
+  width: 100vw;
+}
+
+.bender-center {
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.bender-container {
+  max-width: 600px;
+  width: 100%;
+  padding: 2rem;
+  text-align: center;
+}
+
+.bender-title {
+  font-size: 2.5rem;
+  font-weight: bold;
+  color: #e74c3c;
+  margin-bottom: 1rem;
+}
+
+.bender-subtitle {
+  font-size: 1.2rem;
+  color: #fff;
+  margin-bottom: 2rem;
+}
+
+.bender-image-wrapper {
+  display: flex;
+  justify-content: center;
+  margin-bottom: 2rem;
+}
+
+.bender-image {
+  height: 192px;
+  width: 192px;
+  object-fit: contain;
+}
+
+.bender-quote-bubble {
+  background: #2c3e50;
+  border-radius: 0.4em;
+  padding: 1.5rem;
+  min-height: 150px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-bottom: 2rem;
+  box-shadow: 0 4px 16px rgba(0,0,0,0.2);
+  position: relative;
+}
+
+.bender-quote-bubble:after {
+  content: "";
+  position: absolute;
+  bottom: 0;
+  left: 50%;
+  width: 0;
+  height: 0;
+  border: 20px solid transparent;
+  border-top-color: #2c3e50;
+  border-bottom: 0;
+  border-left: 0;
+  margin-left: -10px;
+  margin-bottom: -20px;
+}
+
+.bender-quote-text {
+  font-size: 1.3rem;
+  color: #fff;
+  width: 100%;
+}
+
+.bender-btn {
+  background-color: #e74c3c;
+  color: #fff;
+  font-weight: bold;
+  padding: 0.75rem 2rem;
+  border: none;
+  border-radius: 999px;
+  font-size: 1.1rem;
+  box-shadow: 0 2px 8px rgba(0,0,0,0.15);
+  transition: background-color 0.3s, transform 0.2s, opacity 0.3s;
+  cursor: pointer;
+  margin-bottom: 1rem;
+}
+
+.bender-btn:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
+}
+
+.bender-btn:not(:disabled):hover {
+  background-color: #c0392b;
+  transform: scale(1.05);
+}
+
+.bender-error {
+  position: fixed;
+  top: 1.25rem;
+  right: 1.25rem;
+  background: #e74c3c;
+  color: #fff;
+  padding: 1rem 1.5rem;
+  border-radius: 0.5rem;
+  box-shadow: 0 2px 8px rgba(0,0,0,0.15);
+  z-index: 1000;
+  transition: opacity 0.3s;
+}
